src/requirements.txt
Django
django-cors-middleware
django-extensions
djangorestframework
PyJWT
six
nose
django_nose
